# [Xamarin.Android.Build.Tasks] remove TFI == MonoAndroid checks