--- a/src/Xamarin.Android.Build.Tasks/Tests/Xamarin.Android.Build.Tests/Tasks/FilterAssembliesTests.cs
+++ b/src/Xamarin.Android.Build.Tasks/Tests/Xamarin.Android.Build.Tests/Tasks/FilterAssembliesTests.cs
@@ -92,7 +92,6 @@ namespace Xamarin.Android.Build.Tests
 			var expected = new [] {
 				"FormsViewGroup.dll",
 				"Xamarin.Forms.Platform.Android.dll",
-				"Xamarin.Forms.Platform.dll",
 			};
 			CollectionAssert.AreEqual (expected, actual);
 		}