@jonpryor 2020-02-06
> We can rely on a reference to `Mono.Android.dll` instead

I would like this to *also* rely on a reference to `Java.Interop.dll`.