@jonathanpeppers 2020-02-06
I made this look for both `Java.Interop.dll` and `Mono.Android.dll` now.