@jonpryor 2020-02-06
I have concerns, as it *looks like* this won't deal with "transitive dependencies".

So consider this silly example:

```bash
$ cat A.cs
using System;
using System.Text.RegularExpressions;

namespace Example {
    public class A {
        Regex r;
    }
}

$ csc /t:Library /nologo A.cs
A.cs(6,15): warning CS0169: The field 'A.r' is never used

$ cat B.cs
using System;

using Example;

namespace Example {
    public class B : A {        
    }
}

$ csc /t:library /nologo B.cs /r:A.dll
```

`A.dll` uses `Regex`, which lives in `System.dll`:

```bash
$ monodis --assemblyref A.dll
AssemblyRef Table
1: Version=4.0.0.0
	Name=mscorlib
	Flags=0x00000000
	Public Key:
0x00000000: B7 7A 5C 56 19 34 E0 89 
2: Version=4.0.0.0
	Name=System
	Flags=0x00000000
	Public Key:
0x00000000: B7 7A 5C 56 19 34 E0 89 
```

`B.dll` references `A.dll`, and type `B` *inherits* `Example.A`, but `B.dll` does *not* reference `System`:

```bash
monodis --assemblyref B.dll
AssemblyRef Table
1: Version=4.0.0.0
	Name=mscorlib
	Flags=0x00000000
	Public Key:
0x00000000: B7 7A 5C 56 19 34 E0 89 
2: Version=0.0.0.0
	Name=A
	Flags=0x00000000
	Zero sized public key
```

Now, why do I care?

Rename `System` to `Mono.Android.dll`, and have `A` be a type which inherits `Java.Lang.Object`, and perhaps you see the issue.  If we have:

```csharp
// Assembly: A.dll
namespace Example {
    public class A : Java.Lang.Object {
    }
}

// Assembly: B.dll
namespace Example {
    public class B : A {
    }
}

// ...and for local testing purposes, Mono.Android.dll:
namespace Java.Lang {
    public class Object {        
    }
}
```

The semantics are that we *must* pass `B.dll` to `<GenerateJavaStubs/>` so that we generate Java Callable Wrappers for it.  Unfortunately, *`B.dll` doesn't reference `Mono.Android.dll`*.

Which is doubly odd, because it *does* need to be referenced:

```
$ csc /nologo /t:library A.cs /r:Mono.Android.dll
$ csc /nologo /t:library B.cs /r:A.dll /r:Mono.Android.dll
$ monodis --assemblyref B.dll
AssemblyRef Table
1: Version=4.0.0.0
	Name=mscorlib
	Flags=0x00000000
	Public Key:
0x00000000: B7 7A 5C 56 19 34 E0 89 
2: Version=0.0.0.0
	Name=A
	Flags=0x00000000
	Zero sized public key
```

Above we see that `B.cs` is compiled into `B.dll`, and *the command line* includes a `/r:Mono.Android.dll`, but *the resulting assembly* does *not* reference `Mono.Android.dll`.

...oops?

---

Maybe this is already handled by the PR; I don't know.  What I'm fairly certain about is that `@(_ResolvedUserMonoAndroidAssemblies)` *must* include *every* assembly that *directly **or indirectly*** references `Mono.Android.dll`, such as the above `B.dll`.

Use of `TargetFrameworkIdentifier` allowed this, as it was auto-inserted into every assembly by MSBuild voodoo.  (Not so much "pure command-line builds", as used above...)  I do not know if this PR makes the same guarantees.