@jonathanpeppers 2020-02-06
> $ monodis --assemblyref B.dll

I need to make a test as mentioned above that verifies the java stub exists after a build. Right now this PR doesn't look like it would.