@jonpryor 2020-02-07
Let's just drop this right here...

![Assembly Dependencies](https://files.slack.com/files-pri/T03C6CW9H-FTBEG36TV/image.png)