@jonathanpeppers 2020-02-07
It definitely breaks in my test that has Foo -> Bar -> Baz:
```
obj\Debug\android\src\crc640a50d1657a1878d1\Foo.java(5,31): error JAVAC0000:  error: package crc647801da6e17af61a6 does not exist
	extends crc647801da6e17af61a6.Bar
 [C:\src\xamarin-android\bin\TestDebug\temp\MissingMonoAndroidReference\FooApp\FooApp.csproj]
```
I almost have it recursively finding the reference for `Bar`, still working on it.