@jonathanpeppers 2020-02-13
I am going to close this for now; I can come back to this if needed.

I was unable to pass the TFI from leaf nodes up to the parent node -- just due to how the ordering/recursive code was working in `<ResolveAssemblies/>`.

What I think we should do instead is make the new .NET 5 TFMs work, when we know what they will be for sure.